feat(auth): defininf the authController and using it
